separated helper functions

+++ routes/user_auth.js
@@ -1,67 +1,17 @@
 const router = require('express').Router()
 require('dotenv').config()
-const jwt = require('jsonwebtoken')
+
+//DBs
 const db = require('../models')
 const { user_auth, user_orders, subscriptions, inventory } = db
+
+//Helper functions
+const { generateAccessToken, decodeToken, doesTokenFail } = require("../helpers/jwt")
+const { noUserIdDuplicates } = require("../helpers/user_db_checks")
 
 //bcrypt
 const bcrypt = require('bcrypt')
 const saltRounds = 10
-
-const userExists = async (user_id)=>{
-    let found = await user_auth.findOne({
-        where: {
-            user_id: user_id
-        }
-    })
-    if(found !== null){
-        return true
-    }
-    return false
-}
-
-const noDuplicates = async (user_id)=>{
-    let user = await user_auth.findOne({
-        where: {
-            user_id: user_id
-        }
-    })
-    if(user){
-        return false
-    }
-    return true
-}
-
-const generateAccessToken = (user_id, email_address, is_admin)=>{
-    let token = jwt
-    .sign(
-        {
-            user_id: user_id, 
-            email_address: email_address,
-            is_admin: is_admin
-        }, 
-        process.env.TOKEN_SECRET, 
-        {
-            expiresIn: process.env.TOKEN_EXPIRATION
-        }
-    )
-    return token
-}
-
-const decodeToken = (req)=>{
-    const token = req.headers.authorization.split(' ')[1];
-
-    if(!token) return new Error("No token found.")
-
-    const decodedToken = jwt.verify(token, process.env.TOKEN_SECRET, (err, decoded)=>{
-        if(err){
-            return
-        }
-        return decoded
-    })
-
-    return decodedToken
-}
 
 //**********ROUTERS*************************************/
 //get all users
@@ -110,18 +60,20 @@ router.post('/', (req, res)=>{
                 last_name: req.body.last_name.toLowerCase(),
                 is_admin: false
             }          
-            if(await noDuplicates(req.body.user_id)){
+            if(await noUserIdDuplicates(user_auth, req.body.user_id)){
                 await user_auth.create(newUser)
-                res.status(200).json({
+                return res.status(200).json({
                     success: true,
                     data: {
                         ...newUser,
                         token: generateAccessToken(newUser.user_id, newUser.email_address, newUser.is_admin)
                     }
                 })
-                return
+                
+            }else{
+                return res.status(203).json({success: false, message: "User already exists."})
             }
-            res.status(203).json({success: false, message: "User already exists."})
+            
         }catch(err){
             res.status(500).json(err)
         }
@@ -167,8 +119,9 @@ router.post('/auth', async (req, res)=>{
 
 //update user information
 router.put('/', (req, res)=>{
-    const token = decodeToken(req)
     
+    const token = decodeToken(req)
+
     if(typeof token === "undefined"){
         return res.status(203).json({
             success: false, 
@@ -176,11 +129,7 @@ router.put('/', (req, res)=>{
         })
     }
 
-    let user_idMatch = req.body.user_id !== token.user_id;
-    let email_addressMatch = req.body.email_address !== token.email_address;
-    let expiredCheck = token.exp >= Date.now();
-
-    if(user_idMatch || email_addressMatch || expiredCheck){
+    if(doesTokenFail(req, token)){
         return res.status(203).json({
             success: false, 
             message: "Authentication error. Invalid bearer token."
@@ -217,27 +166,40 @@ router.put('/', (req, res)=>{
                 }
             })
         }catch(err){
-            return res.status(500).json(err)
+            return res.status(500).json({success: false, error: err, message: "Error updating user."})
         }
     })
 })
 
 //delete user
-//add JWT authentication
 router.delete('/', async (req, res)=>{
+
+    const token = decodeToken(req)
+
+    if(typeof token === "undefined"){
+        return res.status(203).json({
+            success: false, 
+            message: "Authentication error. Invalid bearer token. Error while decoding the token."
+        })
+    }
+
+    if(doesTokenFail(req, token)){ 
+        return res.status(203).json({
+            success: false, 
+            message: "Authentication error. Invalid bearer token."
+        })
+    }
+
     try{
-        user_auth.destroy({
+        await user_auth.destroy({
             where: {
                 user_id: req.body.user_id
             }
         })
-        res.status(200).json("User deleted.")
+        res.status(200).json({success: true, message: "User deleted."})
     }catch(err){
-        res.status(500).json("Error updating user.")
+        res.status(500).json({success: false, error: err, message: "Error deleting user."})
     }
 })
-
-
-
 
 module.exports = router
